+++ CMakelists.txt
@@ -1,20 +1,20 @@
+# cmake最低版本
 cmake_minimum_required(VERSION 3.15)
 
 set(NODE_TARGET x64)
+// 项目名称, 版本号
 project(opengl VERSION 1.0.0)
-
+// c++版本
 set(CMAKE_CXX_STANDARD 11)
 set(CMAKE_CXX_STANDARD_REQUIRED TRUE)
 
-set(BOOST_INCLUDEDIR E:/boost/boost_1_58)
-set(BOOST_LIBRARYDIR E:/boost/boost_1_58/stage/lib64)
 set(OPENGL_INCLUDE F:/OpenGL/include)
 set(OPENGL_LIBRARY F:/OpenGL/libs)
-
+// boost库路径
 find_package(Boost REQUIRED)
+find_package(OpenGL REQUIRED)
 
-
-set(Boost_LIBRARY_DIRS E:/boost/boost_1_58/stage/lib64)
+set(Boost_LIBRARY_DIRS C:/Boost/lib)
 MESSAGE( STATUS "Boost_INCLUDE_DIRS = ${Boost_INCLUDE_DIRS}.")
 MESSAGE( STATUS "Boost_LIBRARIES = ${Boost_LIBRARIES}.")
 MESSAGE( STATUS "Boost_LIB_VERSION = ${Boost_LIB_VERSION}.")
