+++ PreLoad.cmake
@@ -1,1 +1,1 @@
-set (CMAKE_GENERATOR "Visual Studio 15 2017 Win64" CACHE INTERNAL "" FORCE)
+set (CMAKE_GENERATOR "Visual Studio 17 2022 Win64" CACHE INTERNAL "" FORCE)
