--- resource/shader.frag
@@ -1,11 +1,0 @@
-#version 330 core
-out vec4 FragColor;
-
-in vec2 texcoord;
-
-uniform sampler2D texture1;
-uniform sampler2D texture2;
-
-void main(){
-	FragColor = mix(texture(texture1, texcoord), texture(texture2, texcoord), 0.0);
-}
