+++ resource/shader/game.frag
@@ -1,0 +1,7 @@
+#version 330 core
+
+in vec3 fColor;
+
+void main(){
+    FragColor = vec4(fColor, 1.0);
+}
