+++ resource/shader/game.vert
@@ -1,0 +1,9 @@
+#version 330 core
+layout (location = 0) in float aPos;
+
+out vec3 gColor;
+
+void main(){
+    gl_Position = vec4(aPos, 0.5, 0.0, 1.0);
+    gColor = vec3(0.5,0.5,0.5);
+}
