+++ src/header/Shader.h
@@ -9,6 +9,7 @@
 #include <string>
 #include <vector>
 #include "except.h"
+#include "tool.h"
 /*
 .vert - 顶点着色器
 .tesc - 曲面细分控制着色器
@@ -18,13 +19,13 @@
 .comp - 计算着色器
 */
 namespace ShaderFileType {
-    const char * Vert = ".vert";
-    const char * Frag = ".frag";
-    const char * Tesc = ".tesc";
-    const char * Tese = ".tese";
-    const char * Geom = ".geom";
-    const char * Comp = ".comp";
-};  // namespace Type
+extern const char* Vert;
+extern const char* Frag;
+extern const char* Tesc;
+extern const char* Tese;
+extern const char* Geom;
+extern const char* Comp;
+};  // namespace ShaderFileType
 /**
  * @brief shader文件类型
  */
@@ -42,28 +43,28 @@ class Shader {
     std::vector<std::pair<std::string, std::string>> shaderList;
     /**
      * @brief 读取shader文件, 文件类型必须是ShaderType中定义的类型
-     * 
+     *
      * @param path 路径
      */
     void readShader(const std::string& path);
     /**
      * @brief 获取着色器类型对应的OpenGL值
-     * 
+     *
      * @param type 着色器类型
      * @return int 对应的Opengl值
      */
     int getGLShaderType(ShaderType type) const;
 
   public:
-  /**
-   * @brief 构造一个新的Shader对象
-   * 
-   * @param list 着色器文件路径列表
-   */
+    /**
+     * @brief 构造一个新的Shader对象
+     *
+     * @param list 着色器文件路径列表
+     */
     Shader(std::initializer_list<std::string> list);
     /**
      * @brief 构造一个新的Shader对象
-     * 
+     *
      * @tparam T 文件的路径类型--必须能转换为std::string
      * @param path 文件路径
      */
@@ -75,7 +76,7 @@ class Shader {
     }
     /**
      * @brief 构造一个新的Shader对象
-     * 
+     *
      * @tparam T 类型--必须能转换为std::string
      * @tparam Args 类型包--其中每一个必须都能转换为std::string
      * @param path 路径
@@ -89,53 +90,37 @@ class Shader {
     }
     /**
      * @brief Destroy the Shader object
-     * 
+     *
      */
     ~Shader() noexcept {
         if (ID != 0) glDeleteProgram(ID);
     }
     /**
      * @brief 使用着色器
-     * 
+     *
      */
     void use();
     /**
      * @brief 设置着色器中的Uniform对象
-     * 
+     *
      * @param name Uniform对象名称
      * @param value 需要设置的int值
      */
     void setUniform(const std::string& name, bool value) const;
     /**
      * @brief 设置着色器中的Uniform对象
-     * 
+     *
      * @param name Uniform对象名称
      * @param value 需要设置的int值
      */
     void setUniform(const std::string& name, int value) const;
     /**
      * @brief 设置着色器中的Uniform对象
-     * 
+     *
      * @param name Uniform对象名称
      * @param value 需要设置的float值
      */
     void setUniform(const std::string& name, float value) const;
-    /**
-     * @brief 设置着色器中的Uniform对象
-     * 
-     * @param name Uniform对象名称
-     * @param val 需要设置的mat4值
-     * @param count 需要加载数据的数组元素的数量或者需要修改的矩阵的数量--默认为1
-     */
-    void setUniform(const std::string& name, glm::mat4 val,
-                    int count = 1) const;
-    /**
-     * @brief 设置着色器中的Uniform对象
-     * 
-     * @param name Uniform对象名称
-     * @param val 需要设置的vec3值
-     */
-    void setUniform(const std::string& name, glm::vec3 val) const;
     /**
      * @brief 编译shader
      *
@@ -155,11 +140,30 @@ class Shader {
     unsigned int id() const { return ID; }
     /**
      * @brief 获取着色器文件对应的ShaderType
-     * 
+     *
      * @param path 着色器文件的路径
      * @return ShaderType 对应的ShaderType
      */
     ShaderType getShaderType(const std::string& path) const;
+    /**
+     * @brief 设置着色器中的Uniform对象
+     *
+     * @tparam 模板参数
+     * @param name Uniform对象名称
+     * @param val 需要设置的值
+     */
+    template <class Vector>
+    void setUniformV(const std::string& name, Vector val) const;
+    /**
+     * @brief 设置着色器中的Uniform对象
+     *
+     * @tparam Matrix
+     * @param name Uniform对象名称
+     * @param val 需要设置的值
+     * @param count 需要设置的数量默认1个
+     */
+    template <class Matrix, int N = 1>
+    void setUniformM(const std::string& name, Matrix val, int count = N) const;
 };
 
 #endif
