+++ src/header/config.h
@@ -1,5 +1,5 @@
 #pragma once
-
+/*
 #ifndef DLL_API
 #ifdef DLL
 #define DLL_API __declspec(dllexport)
@@ -7,5 +7,4 @@
 #define DLL_API __declspec(dllimport)
 #endif
 #endif
-
-#define PI 3.1415926
+*/
