+++ src/header/tool.h
@@ -1,56 +1,7 @@
 #pragma once
-#include <glad/glad.h>
-#include <GLFW/glfw3.h>
-#include "stb_image.h"
+
 #include <iostream>
 #include "utils.h"
-
-/**
- * @brief 初始化OpenGL环境
- * 
- * @return GLFWwindow* 指向GLWFwindow的指针
- */
-GLFWwindow* init();
-/**
- * @brief 键盘回调函数
- * 
- * @param window 指向GLWFwindow的指针
- */
-void processInput(GLFWwindow* window);
-/**
- * @brief 读取纹理图片
- * 
- * @param path 纹理图片路径
- */
-void readImage(const char* path);
-/**
- * @brief 载入纹理图片
- * 
- * @param path 纹理图片路径
- * @return unsigned int 纹理名称
- */
-unsigned int loadTexture(const char* path);
-/**
- * @brief 画面调整回调函数
- * 
- * @param window 指向GLWFwindow的指针
- * @param width 宽度
- * @param height 高度
- */
-void framebuffer_size_callback(GLFWwindow* window, int width, int height);
-/**
- * @brief 加载天空盒
- * 顺序:
- * +X (right)
- * -X (left)
- * +Y (top)
- * -Y (bottom)
- * +Z (front)
- * -Z (back)
- * @param faces 包含6个纹理路径的数组
- * @return unsigned int 纹理名称
- */
-unsigned int loadCubemap(std::vector<std::string> faces);
 
 const float cubeVertices[] = {
     // positions          // texture Coords
